css: dark theme: fix panel borders for pmg eol theme
since the header markup is different than what we use in pve, the
'spacer' elements in the pmg header get a (different colored) border by
default.

We already overwrite it for the first spacer, but when we show an EOL
notice, we have two of them, so add the second one to the rule too.
This also uses the 'css next sibling selector', just for the sibling of
the div with the 'eol-notice' class instead of the 'versioninfo' one.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>