object grid: fix onlineHelp setting from editorConfig for row editors
In our row editors helpers, we unconditionally set onlineHelp from
'opts.onlineHelp', even if it's undefined.

Later we use 'Ext.apply' to set first the editorConfig defaults, then
the 'rowdef.editor' settings. In javascript, the objects

{} and
{ foo: undefined }

are not the same, so Ext.apply overwrites the default from editorConfig
with that from the row definition, also for undefined.

This means if we have a default onlineHelp in editorConfig and none in
the add_*_row options, we would not show it.

To fix it, check if 'opts.onlineHelp' is truthy before setting it in
the row definition. This should not happen for other options used
from the row helper options, since those are nested
(Ext.apply does not work recursively)

This fixes a regression in pmg-gui, where we set a default onlineHelp
for e.g. the Mail Proxy Options which would not show up anymore.

Note: PMG is the only product where we used this pattern, so this
was not visible anywhere in PVE or PBS.

Fixes: 7d16f8b (object grid: allow to pass online help to row editors)
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-By: Stoiko Ivanov <s.ivanov@proxmox.com>
Tested-By: Stoiko Ivanov <s.ivanov@proxmox.com>