fix #6312: limit consent window to viewport size
The consent window will try to size itself according to the max/min
constraints set, but those might be too large for some viewport sizes.

Since it's not possible to set those to relative viewport sizes (ExtJS
does it's own layout, so we can't use CSS), we have to save the original
size and on resize of the viewport limit ourselves to that (minus 10 px
safety margin).

Even with that the window will adhere to the set constraint sizes,
but it should be viewable with any sensible display/viewport size.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250515080214.734035-1-d.csapak@proxmox.com
 [TL: re-format with proxmox-biome]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
(cherry picked from commit 5b329a62352a14544b313fd5762d782efa480231)
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>