notification: smtp: only show 'Unchanged' in password field when it makes sense
The formula used for deriving the emptyText for the password field did
not correctly pick up the 'isCreate' flag, so 'Unchanged' was also
visible when creating a new SMTP target that has no password yet.

Also took the opportunity to improve the behavior when modifying an
existing target. Now the original state of the 'authentication' flag is
saved. The 'Unchanged' emptyText is only shown if the loaded config
contains authentication parameters. When editing a target *without*
authentication, the password field does not have the 'Unchanged'
emptyText any more.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Michael Köppl <m.koeppl@proxmox.com>
Reviewed-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20250627121339.182283-2-l.wagner@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>