node service view: increase log window size depending on viewport
The 2:1 ration of the 800x400 size looked rather odd and cramped,
instead of just increasing the height default to different sizes
depending on the viewport size (at time of opening).

We really should start a common helper for this which returns
two or three different sizes depending on the view port size so that
we can move most windows that can profit from being bigger, like task
log, over to it and have unified behavior.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>