network selector: add type parameter
The network endpoint of the PVE API allows selecting interfaces by
type, but the network selector widget currently does not support
passing the type to the API call.

This is required for the SDN fabrics, which introduced a new special
type to this endpoint that additionally selects all SDN interfaces.
This can then be used to make SDN fabrics show up in the Migration
Settings dialog or in the Ceph Installation Wizards.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-57-g.goller@proxmox.com