rrdchart: allow to override the series object
this way we can keep the current behavior, but also make it possible to
finely control a series if needed. For example, if we want a stacked
graph, or just a line without fill.

Additionally we need to adjust the tooltip renderer to also gather the
titles from these directly configured series.
We also add a check if the value is null and set the tooltip to "No
Data". Because stacked graphs (which we want to use now), will graph
those data points as well.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250726010626.1496866-7-a.lauterer@proxmox.com