rrdchart: add dummy listener for legend clicks
This way we can define a listener when needed to react to any clicks in
the legend. Usually enabling/disabling some data series.

The event was nowhere documented, but by using the following snippet,
right where we add the listener, it can be observed to happen.

```
Ext.mixin.Observable.capture(this, function(evname) {console.log(evname, arguments);})
```
This would be `me.legend` in that scenario.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Link: https://lore.proxmox.com/20250726010626.1496866-10-a.lauterer@proxmox.com