network edit: fix hiding of altnames field if disabled by config
Commit 691f530d ("network: optionally show alternative interface
names") introduced the optional altnames field in the network edit
grid. The chosen default configuration however does not even add the
field, therefore setting it to hidden if no altnames data is present
cannot work.

Fix this by only hiding the field if it was actually configured to be
shown in the first place.

Fixes: 691f530d ("network: optionally show alternative interface names")
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250731113015.148881-1-c.ebner@proxmox.com