add Proxmox logo as SVG
The SVG is based on the one from PDM, but with the `width` and `height`
properties removed - apparently that is needed so that the SVG properly
resizes with the parent container. Also the media query which changes
the fill color for the dark theme support was removed - since that
happens on the CSS level now, same as for old PNG logo.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20250804090228.192088-2-l.wagner@proxmox.com