apt repositoires: allow major upgrade
Same as 622ba1f ("apt repositoires: allow major upgrade") in the
stable-bullseye branch.

Currently, all this does is silence warnings when the suite after the
current one is configured as well as enabling the check for mixed
repositiories.

Since there is no API call as mentioned in the comment yet, just set
it manually.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250804132458.100337-1-f.ebner@proxmox.com