RRDChart: fix title lookup for object series
when we give a field(series) as an object instead of a string, we have
to lookup the title from the object. The code was designed to do this,
but accidentally always used the first entry in the 'fields' array
to lookup the title if the field was not a plain string one.

to fix this, simply check if the field is the one we're looking for

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250804132928.1714525-1-d.csapak@proxmox.com