RRDChart: fix customizing axes in a subclass
the introduction of the NumericBase2 segmenter in commit 9531c65 (rrd
chart: fix y-axis segmentation when using powerOfTwo) changed the way
the axes are defined and created. The commit message noted that we
can't override the axes config anymore but that we did not use that.
It actually was used in the PMG dashboard with the 'MiniGraph'
subclass. Luckily the configs were similar enough so that it did not
matter much, so nobody seemed to notice.

On recent examination of the charts, the labels of the axes were
overlapping each other in certain conditions, and then it became
apparent that the customization there for the axis did not apply
anymore.

To fix it, move the axes definition back to the class config, and
overwrite the segmenter on each left or right 'numeric' axis if
necessary.

By overwriting it in 'config.axes' (and me.config.axes as fallback)
we're able to use:
* a default config (from the class definition, lands in `me.config`)
* a subclass override (lands also in `me.config`)
* overrides on actual creation (lands in `config`)

I examined if we can override/set this config later e.g. in the
initComponent, since a manual override + class config would be merged
there, but it seems that the chart initializes the axes already before
the initComponent.
And using 'setSegmenter' later on is not really possible, because for
some reason this only changes the top-most point in the axis labels,
but not the intermediate ones.

Reported-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250919080358.745214-1-d.csapak@proxmox.com
 [TL: tiny reword/reflow of commit message and record R-b]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>