window: introduce dangerous parameter to SafeDestroy
The 'dangerous' parameter determines whether the user will be required to
enter the ID of the item they want to delete to confirm their intent. If
set to true, the input field will be added and a warning icon will be
displayed. If false, there is no ID input field and a question mark will
be displayed.

Signed-off-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20250930145848.263162-7-m.koeppl@proxmox.com
Reviewed-by: Daniel Kral <d.kral@proxmox.com>
Tested-by: Daniel Kral <d.kral@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>