window: acme account: hide e-mail field if it does not exist
Let's Encrypt stopped sending out email notifications for upcoming
certificate expiries in June 2025 [0], and with that they seem to have
stopped returning the contact field in the account information
anymore.

To not trip up the ui, only optionally read that out, and hide the field
completely if it does not exist. This should keep compatibility for
other ACME providers that still have a contact field.

[0]: https://letsencrypt.org/2025/01/22/ending-expiration-emails

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251002112528.3128880-1-d.csapak@proxmox.com
 [TL: provide some more context in commit message]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>