fix #6830: ui: apiviewer: add description panel to return section
The PVE API viewer does not display schema object descriptions.
For example, when an endpoint's return type is defined like this:

`returns => {
    type => 'object',
    description => 'Some Description.',
    ...
}`

the description field would not be shown anywhere in the API viewer.

This patch adds a panel above the properties table, which shows the
description of the object if it is set.

Fixes: https://bugzilla.proxmox.com/show_bug.cgi?id=6830
Signed-off-by: Nicolas Frey <n.frey@proxmox.com>
Link: https://lore.proxmox.com/20251015140350.176989-1-n.frey@proxmox.com