rrd metrics: split time-frame and aggregation-type into separate selectors
Instead of duplicating each time-frame for both available aggregation
types, split the existing combobox into one that just provides the
'timeframe' selector and a segmented button with the aggregation types
(average and maximum), i.e. what RRD calls the consolidation function.

This removes some clutter in the dropdown and allows for a slightly
more convenient and intuitive UX, as e.g. one now can quickly switch
between average and maximum data points for the same time period.

This approach was originally developed for PDM in its Yew (rust) based
UI, but it's easy enough to "port" that change over to the ExtJS based
UIs.

Suggested-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251107140400.4016493-1-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>