rrd metrics: make segmented buttons of aggregation-type selector less flashy
The two blue "sub" buttons where rather flashy here and drew more
attention as warranted, so rather use the inline button style.

For our existing inline button style to look OK for the two segmented
buttons, we need to clear the border radii for the corners in the
middle where the two buttons meet, and set the border width to zero
for the site where they meet.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>