pending object grid: don't fall back to empty string for 0 values
if we have a value that is set to `0` (the number, not the string),
the default pending object grid renderer would overwrite that with the
empty string, since `0` does not evaluate to true.

Instead, use the `Nullish coalescing operator` (??) to overwrite the
value with only when it's null or undefined.

This fixes an issue with the rendering if the `TTY Count` in PVE is set
to 0. (There are no other options where this could happen currently,
because they either have a custom renderer, or do not allow 0 as value).

Reported-by: Dietmar Maurer <dietmar@proxmox.com>
Suggested-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251117112213.1387848-1-d.csapak@proxmox.com