combo grid: record todo for renaming skipEmptyText to something less confusing
emptyText is a wide-spread option in ExtJS which is used to set the
placeholder text that is visible when a field is empty.

The skipEmptyText option from our ComboGrid widget OTOH is completely
unrelated to that emptyText option, but rather defines if an empty
string should be submitted or mapped to null (which is normally
filtered out from submission).

Record that discrepancy in a comment as I do not got the time to fix
this backward-compatible manner just now, and besides serving as a
reminder the comment  might already help some that try to look into
what this option does.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>