RRD metric type selector: adapt applying CSS class to fix ExtJS debug mode
For some reason, using a space separated list of classes in `itemCls`
breaks the extjs debug mode. This is seemingly allowed everywhere
else, so to fix it, simply apply the classes directly to the children
with 'default'.

An alternative would be to switch itemCls over to an array, but using
a default here is actually a bit more telling and more importantly,
avoids accessing some internal property that isn't really documented.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251205143450.2953331-1-d.csapak@proxmox.com
 [TL: add some more context from my time running into this]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>