fix #7087: ui: enable direct task log download
The web ui of both, PVE and PBS, now includes a "Download" button
allowing for a task log to be downloaded directly without opening the
task viewer pop-up.

The enhancement originally related to PBS only, however this change
applies to PVE too and it seems more convenient to have it for both.

Signed-off-by: Dominik Rusovac <d.rusovac@proxmox.com>
Link: https://lore.proxmox.com/all/20251219144635.168818-1-d.rusovac@proxmox.com